backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
numpy==1.24.3
Pillow==10.0.0
pymongo==4.6.0
dnspython==2.4.2
gunicorn==21.2.0
tflite-runtime==2.14.0
